DBSCAN code move, notebook, benchmark (#608)
* DBSCAN notebook, benchmark script

* Move DBSCAN to clustering


* File check in

* benchmark script support

* Benchmark parameter fix

* Transform time fix

* DBSCAN data broadcast fix, comment fix

* separate change for DBSCAN source code fix

* cmdline switch for benchmark score compute

* style fix

---------
